requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-docx==0.8.11
PyPDF2==3.0.1
Werkzeug==2.3.7
gTTS==2.4.0
pyttsx3==2.90
pygame==2.5.2
